- (keytools.c, keytools.h, parse.c, scapi.c): Patch from Markku Laukkanen:   - size_t changes.
git-svn-id: file:///home/hardaker/lib/sf-bkups/net-snmp-convert-svnrepo/trunk@2885 06827809-a52a-0410-b366-d66718629ded

+++ snmplib/keytools.h
@@ -37,14 +37,14 @@ extern "C" {
 extern int (*kmt_hash) (
 	const int	  mode,		void  	 **context,
 	const u_int8_t	 *data,		const int  data_len,     
-	u_int8_t	**digest,	int	  *digest_len);
+	u_int8_t	**digest,	size_t	  *digest_len);
 
 
 #ifdef HAVE_LIBKMT
 extern int (*kmt_s_md5) (
 		const int	  mode,		void  	 **context,
 		const u_int8_t	 *data,		const int  data_len,     
-		u_int8_t	**digest,	int	  *digest_len);
+		u_int8_t	**digest,	size_t	  *digest_len);
 extern int (*kmt_s_sha1) (
 		const int	  mode,		void  	 **context,
 		const u_int8_t	 *data,		const int  data_len,     
@@ -52,7 +52,7 @@ extern int (*kmt_s_sha1) (
 extern int (*kmt_s_ripemd) (
 		const int	  mode,		void  	 **context,
 		const u_int8_t	 *data,		const int  data_len,     
-		u_int8_t	**digest,	int	  *digest_len);
+		u_int8_t	**digest,	size_t	  *digest_len);
 
 #endif
